requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
docling
docetl
anthropic
python-multipart
